feat: the default provider now handles major versions
Before the JBang provider was responsible for handling the mapping of
major versions to the JDKs that are installed. But now that it can install
multiple JDKs with the same major version, the responsibility for creating
the mapping has been moved to the default provider. It will manage symlinks
from major version numbers to JDKs.

Fixed #47